requirements.txt
langgraph
langchain
langchain-openai
openai
replicate
fastapi
uvicorn
python-dotenv